+++ src/main/java/com/github/lmh01/mgt2mt/content/manager/DevLegendsManager.java
@@ -2,10 +2,8 @@ package com.github.lmh01.mgt2mt.content.manager;
 
 import com.github.lmh01.mgt2mt.MadGamesTycoon2ModTool;
 import com.github.lmh01.mgt2mt.content.DevLegend;
-import com.github.lmh01.mgt2mt.content.Licence;
 import com.github.lmh01.mgt2mt.content.managed.*;
 import com.github.lmh01.mgt2mt.content.managed.types.DevLegendType;
-import com.github.lmh01.mgt2mt.content.managed.types.LicenceType;
 import com.github.lmh01.mgt2mt.util.I18n;
 import com.github.lmh01.mgt2mt.util.MGT2Paths;
 import com.github.lmh01.mgt2mt.util.Utils;