+++ changelog.md
@@ -8,6 +8,7 @@
 ### Other
 - When a new update is available and the GitHub repository is opened, the specific release page is now opened
 - Optimized localisation files
+- Error message is now written to the console when the import fails because image files could not be copied
 
 ### Buf fixes
 - Fixed error message formatting when restore point content could not be constructed